Dont check ignores if in blocking func
For the case of MPI on NSCC cluster, it calls into libfabric.so which
ends up in the following scenario.

- libfabric calls __interceptor_poll (in TSan)
- TSan calls ScopedInterceptor constructor
  - thr->ignore_reads_and_writes++
- TSan calls real poll in libc

Supposedly after returning from real poll, TSan should call the
ScopedInterceptor destructor, which does
  - thr->ignore_reads_and_writes--

However, real poll in libc somehow calls __nptl_deallocate_tsd.part.8
in /lib64/libpthread.so.0. This is for deallocating the TLS.

Somehow __nptl_deallocate_tsd.part.8 calls ThreadFinish.

As thr->ignore_reads_and_writes did not get decremented as expected,
ThreadCheckIgnore will call ReportIgnoresEnabled which calls Die(),
killing the whole process and all the threads in it.

---

The root cause is that TSan did not properly return from a blocking
function call and thus did not call ~ScopedInterceptor that will
decrement thr->ignore_reads_and_writes as it should have done.
Instead, it calls ThreadFinish without doing that.

As such, we should not be checking for ignores if TSan happens to
call ThreadFinish while still inside a blocking function call.

This should not be an issue wrt the number of warnings reported
(no FP) because there should not have been any races detected when
this happens.

---

Reproduce: AMG benchmark running on NSCC normal node, compiled with
CCE cc, with -fno-cray flag