[psan] some fixes
This includes a hacky fix that I don't really like for the
signal_thread2.cpp test which fails due to both segfault and
deadlock.

There are other miscellaneous fixes in this commit that is not
related to the bug.

Ideally some time later we should have our own allocator for
shadow objects for better performance. But this might be a major
effort in itself.