[psan] better memory usage
ShadowSetWrite only sets the fast epoch and does not load/create
a HBShadowCell. This is important for both speed and memory usage.
The old naive way of loading HBShadowCell for every byte does not
work at all when there are many allocations by the underlying
program.

Now all tests pass except for large_malloc_meta which we will ignore
for now.

TODO:
- revisit all the reinterpret_casts and make sure they are used
  correctly
- reclaim HBShadowCells when underlying memory is freed
- check large_malloc_meta test case